requirements.txt
langchain_core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
faiss-cpu
groq
langchain-groq
wikipedia